Network persistent info (last IP + client id); commit hooks to correctly apply combine changes
Now e.g. addrs + routes are applied together at the end of transaction, thus
we can do the right thing instead of relying on the order in which Rulebook
sets individual attributes (if the order was wrong, we would try e.g. to
set default gateway before we have an address, resulting in a "Network unreachable"
error from ``ip route add``).